backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-search-documents
pypdf
Pillow


